fixes pathing issues in assignment
