+++ src/pointer.rs
@@ -560,7 +560,7 @@ impl PointerBuf {
     }
 
     /// Creates a new `PointerBuf` with the given capacity.
-    pub fn with_capacity(capacity: usize) -> Self {
+    pub(crate) fn with_capacity(capacity: usize) -> Self {
         Self(String::with_capacity(capacity))
     }
     /// Creates a new `PointerBuf` from a slice of non-encoded strings.
@@ -1125,174 +1125,6 @@ mod tests {
         let err = res.unwrap_err();
         assert!(err.is_unreachable());
         assert_eq!(err.offset(), 9);
-    }
-
-    #[test]
-    fn test_assign() {
-        let mut data = json!({});
-        let ptr = Pointer::from_static("/foo");
-        let val = json!("bar");
-
-        let assignment = ptr.assign(&mut data, val.clone()).unwrap();
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(assignment.assigned, &val);
-        assert_eq!(assignment.assigned_to, "/foo");
-
-        // now testing replacement
-        let val2 = json!("baz");
-        let assignment = ptr.assign(&mut data, val2.clone()).unwrap();
-        assert_eq!(assignment.replaced, Some(Value::String("bar".to_string())));
-        assert_eq!(assignment.assigned, &val2);
-        assert_eq!(assignment.assigned_to, "/foo");
-    }
-
-    #[test]
-    fn test_assign_with_explicit_array_path() {
-        let mut data = json!({});
-        let ptr = Pointer::from_static("/foo/0/bar");
-        let val = json!("baz");
-
-        let assignment = ptr.assign(&mut data, val).unwrap();
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(assignment.assigned_to, "/foo/0/bar");
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(
-            json!({
-                "foo": [
-                    {
-                        "bar": "baz"
-                    }
-                ]
-            }),
-            data.clone()
-        );
-    }
-
-    #[test]
-    fn test_assign_array_with_next_token() {
-        let mut data = json!({});
-        let ptr = PointerBuf::try_from("/foo/-/bar").unwrap();
-        let val = json!("baz");
-        let assignment = ptr.assign(&mut data, val).unwrap();
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(
-            assignment.assigned_to, "/foo/0/bar",
-            "`assigned_to` should equal \"/foo/0/bar\""
-        );
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(
-            json!({
-                "foo": [
-                    {
-                        "bar": "baz"
-                    }
-                ]
-            }),
-            data.clone()
-        );
-        let val = json!("baz2");
-        let assignment = ptr.assign(&mut data, val).unwrap();
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(
-            assignment.assigned_to, "/foo/1/bar",
-            "`assigned_to` should equal \"/foo/1/bar\""
-        );
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(
-            json!({
-                "foo": [
-                    {
-                        "bar": "baz"
-                    },
-                    {
-                        "bar": "baz2"
-                    }
-                ]
-            }),
-            data.clone()
-        );
-        let ptr = PointerBuf::try_from("/foo/0/bar").unwrap();
-        let val = json!("qux");
-        let assignment = ptr.assign(&mut data, val).unwrap();
-        // assert_eq!(assignment.assigned_to, "/foo/0/bar");
-        assert_eq!(assignment.replaced, Some(Value::String("baz".to_string())));
-        assert_eq!(
-            json!({
-                "foo": [
-                    {
-                        "bar": "qux"
-                    },
-                    {
-                        "bar": "baz2"
-                    }
-                ]
-            }),
-            data.clone()
-        );
-    }
-
-    #[test]
-    fn test_assign_with_obj_path() {
-        let mut data = json!({});
-        let ptr = PointerBuf::try_from("/foo/bar").unwrap();
-        let val = json!("baz");
-
-        let assignment = ptr.assign(&mut data, val).unwrap();
-        assert_eq!(assignment.assigned_to, "/foo/bar");
-        assert_eq!(assignment.replaced, None);
-        assert_eq!(
-            json!({
-                "foo": {
-                    "bar": "baz"
-                }
-            }),
-            data.clone()
-        );
-    }
-
-    #[test]
-    fn test_assign_with_scalar_replace() {
-        let mut data = json!({
-            "foo": "bar"
-        });
-
-        let ptr = Pointer::from_static("/foo/bar/baz");
-        let val = json!("qux");
-
-        ptr.assign(&mut data, val).unwrap();
-        assert_eq!(
-            &json!({
-                "foo": {
-                    "bar": {
-                        "baz": "qux"
-                    }
-                }
-            }),
-            &data
-        );
-    }
-
-    #[test]
-    fn nested_maps_with_empty_keys() {
-        let data = json!({
-            "": {
-                "": {
-                    "bar": 42,
-                }
-            }
-        });
-
-        {
-            let ptr = Pointer::from_static("///bar");
-            assert_eq!(ptr.resolve(&data).unwrap(), 42);
-        }
-        {
-            let mut ptr = PointerBuf::new();
-            ptr.push_back("".into());
-            ptr.push_back("".into());
-            ptr.push_back("bar".into());
-            assert_eq!(ptr.resolve(&data).unwrap(), 42);
-        }
     }
 
     #[test]
